Ensure Gradle can run without passing the snapshot repo password (#13)
The project property 'snapshotRepoPass' needs to be defined for the script
to run successfully. To make sure it does not have to be specified as a
command line argument for _every_ Gradle run (but only for those that
publish artifacts) the property is defined empty if it was not given as a
command line argument.