Update package names to new project name (#24, #45)
With the project being renamed from "Junit Io" to "JUnit Pioneer" the
old top level package name `org.codefx.junit.io` had to be changed.
Because the project owns the domain junit-pioneer.org, it was obvious
to use it as a package name according to the establishes reverse DNS
naming strategy. In that case the dash (-) has to be replaced with a
valid Java package identifier character. In a bold move (we're
pioneers, after all!) the dollar sign ($) was chosen. :)