Read Maven Central credentials from Gradle properties (#46 / #47)
To allow all maintainers to publish to Maven Central the credentials
are read from the Gradle properties file. This did not actually
require a technical change as the existing mechanism already did that
implicitly (and without my knowledge). It is also still possible to
specify credentials on the command line with `-P property=value`. New
is that the user name as well as the password is read from properties
and that CONTRIBUTING.md describes the process.

This required changes to CI. Travis configuration `.travis.yml` was
updated to pass values for user name and password on the command
line.