Document adaptation to upstream changes (#30 / #48)
The development process regarding branching, commits, pull requests,
and merging defined in #40 is fine for bug-fixing and feature
development but incurs a high overhead for the frequent changes that
need to be made due to Pioneer's close coupling to JUnit 5. For such
cases, a new set of guidelines was created that make those changes
easier.