Create blueprints for project site and documentation (#11, #19 / #54)
The bulk of the work for a project site is done in [junit-pioneer.github.io](https://github.com/junit-pioneer/junit-pioneer.github.io) but it relies on project-specific files to fill the site and particularly the documentation with content. This PR shows what such files would look like:

* pages and documentation can be written in Asciidoc
* a sidebar menu for the documentation is enabled and must be maintained manually
* a project page and documentation page are required

The work in the .io repository makes sure that these files are automatically put in the right places, so the effort within this project should be fairly small. It would be nice to get feedback on whether the features provided here seem reasonable or lacking something that is needed from the start.

Relates to #11 and #19 but closes neither.