Define automatic module name in manifest (#60 / #61)
With Java 9 on the doorstep, JUnit Pioneer needs to be prepared to be
used as a module. Until a proper modularization includes a module
descriptor in the JAR, it is important to define its future module
name. This way, when used as an automatic module that other modules
depend on, releasing a modular JAR does not break those modules.