Introduce TempDirectory Jupiter Extension (#39)
The `TempDirectory` extension can be used to create and clean up a
temporary directory for an individual test or all tests in a test class.
To use it, simply register the extension and add a parameter of type
`java.nio.file.Path` to your test or lifecycle method or constructor:

  @ExtendWith(TempDirectory.class)
  class Tests {
    @Test
    void test(@TempDir Path tempDir) throws Exception {
      Path file = tempDir.resolve("test.txt");
      writeFile(file);
      assertExpectedFileContent(file);
    }
  }

In addition to the default file system, the extension can be used with
any FileSystem implementation, e.g. Jimfs [1]. In order to use a custom
file system, simply register the extension programmatically and pass a
provider of a custom parent directory of type `Path` to its constructor.

[1] https://github.com/google/jimfs