Make DefaultLocale configurable via language tag
This adds a `value()` setting to the `@DefaultLocale` annotation. If the value is set it will be used
to create the default locale via the `Locale.forLanguageTag(String)` method. To make this 
possible, the `language()` setting is no longer mandatory. The language tag based configuration and
the old configuration can't be mixed: If `value()` is set, any other setting will cause an exception to be
thrown. Existing code will still compile and work the same as before.

Closes: #114
PR: #128