Reactivate site build with more verbose error messages
The site build could not be triggered, apparently because the CI
setup[1] didn't set the environment variable `TRAVIS_API_TOKEN`.
We must have accidentally deelted it. I added it back and this
commit reactivates the deployment step.

The class `TriggerTravisTask` now checks whether the token is
empty and prints an informative message before exiting if it is.
This is more informative than a 403 from the Travis endpoint.

Closes: #124
PR: #132

[1] https://travis-ci.org/junit-pioneer/junit-pioneer/settings