Generalize the RangeSourceProvider
This patch generalizes the RangeSourceProvider to allow multiple range
source annotations for different primitive ranges.
The current patch just refactors the code to allow adding additional
annotations in the subsequent patches, but does not add any
functionality.

RangeSourceProvider is refactored so it is no longer an
AnnotationConsumer, but instead extracts the values from the
annotation in its provideArguments method.
This extraction is based on naming conventions of the annotation's
properties, which must be as follows:
- step - a numeric primitive, cannot be zero
- from - a numeric primitive
- to - a numeric primitive
- closed - a boolean
step, from and to must have the same type.

RangeSource is renamed to IntRangeSource, to highlight its purpose.