Implement @ClearSystemProperty and @SetSystemProperty
Sometimes it is helpful to clear and/or set system properties for a
test execution. This PR introduces two annotations,
`@ClearSystemProperty` and `@SetSystemProperty`, that allow to do
this in a save manner with a proper restore mechanism. Both
annotations work on the test method and class level, are repeatable
as well as combinable.

Fixes: #129
PR: #133

[ci skip-release]