With this changes we are able to track code changes with sonarcloud.io.
It provides configruable Quality Gates and Profiles to be applied to
the analysis to maintain a healthy code base. It includes checks for
coverage and duplication. Additionally it will be possible to add
further analysis tools, as soon as they are available.

There are some little hacks like parsing the Sonarqube_params from the
travisCI addon, to ensure pullrequest analysis.

- adding jacoco for coverage
- adding sonarqube to build.gradle
- adding advanced handling of travis ci params to build.gradle
- adding build steps to travisCI