Moving to Github Actions
As Github offers now its own CI Tool with GithubActions,
we want to migrate to this new approach, as it is tightly
coupled with github anyway.

Sonarqube injects its proeprties already correctly, so it
should be fine to use.

We only have to add the SONAR_TOKEN to the project secrets,
but that is nothing i can do.

Issue: #192