GitHub Action for Deployments
Adding GitHub Actions for Deployment which will run
on Linux and with JDK 8. On the side i removed a lot
of unnecessary gradle code for publishing.

To get this running within the base repository, we
need to add following secrets:

GH_WRITE_TOKEN: for updating the repo and setting the tag
BINTRAY_USER_NAME: for upload to bintray
BINTRAY_API_KEY: for upload to bintray
NEXUS_TOKEN_USERNAME: for syncing bintray to maven central
NEXUS_TOKEN_PASSWORD: for syncing bintray to maven central