Configure GitHub Action for deployments
Add GitHub Actions for deployment, which will run on Linux and with
JDK 8. This step allows us to remove Travis completely from the
project. Additionally, we also remove AppVeyor as it is now also
redundant with GitHub Actions.

As the Sitebuild was broken, we also removed this implementation and
we will, in the next step, start implementing a new SiteBuild with
GitHub Actions.

Closes: #199, #234
Relates: #192
PR: #213