Trigger Release with tag 'releaseTrigger'
This changes will trigger a build on master,
when there is a tag 'releaseTrigger' created.

It will notify our normal GitHub Actions, and
retrigger a build. The payload will tell the
our normal build, that this is a release, and
it will trigger the build.

Downside, this will only work in MASTER for now.
If we want to support other branches, we need
to use our combined brain power, how to retrieve
a branch name from a git tag. It is not such
a trivial problem.