Replace automatic releases with manual trigger
We used Shipkit for a continuous delivery model, where every change
on the `master` branch (for example when merging a pull request)
triggered a release build that published a new version if specific
criteria were met. Because this project's development often happens
in sporadic bursts, where a lot of PRs are merged within a few hours,
this approach lead to some superfluous releases. We also weren't 100%
successful in predicting whether Shipkit would make a release and so
we started cluttering our commit messages with `[ci skip-release]`,
which was a bit annoying. Hence the change to a new release model.

With this change, we no longer release on every commit to `master`.
Instead, releases must be triggered manually by pushing a tag
`releaseTrigger` to `master`. GitHub Actions will then tell Shipkit
to do its thing and afterwards delete the tag.

See issue #193 for a discussion of other option to trigger a release.

This change also triggers a site build on successful release, which
finishes the move from Travis CI to GitHub actions.

Closes: #193, #192
PR: #256