Creating IDE Specific configuration with gradle
I added a small build.gradle file, which uses gradle dsl, as i found it
easiert to manipulate xml with. This snippet will add the issue link to
idea on each sync.

The questions to answer is, if we want to stick to groovy and mix kotlin
and groovy dsl. Which might be a bad idea. Sadly i have not had the time
to check if there is a convenient way to do so.

I just want to show this, and cross check if that might be suitable for
our cause or not. (i think at least for now it is)

closes: #229