Rename `RepeatFailedTest` to `RetryingTest` (#233 / #236)
#233 reported a problem that was eventually tracked down to
`@RepeatFailedTest` and `@Test` being both applied to the same
method. While the extension's documentation clearly advises against
that, we think the name contributed to the confusion:
`@RepeatFailedTest` may sound too much like needing to be applied to
a method that is already a test, i.e. already annotated with `@Test`.
We hope that `@RetryingTest` prevents that.

Closes: #233 
PR: #236