Improving and fixing releasing process (#284)
Releasing process was failing, as it could not checkout the ref
'master'. With this changes we bound the releasing process to
the master branch, as we are now checking out the master
specifically.

Furthermore we found issues, with some orders in the releasing
process, which could prevent the deletion of the tag on a failure.

Additionally we added a new last step to our normal build pipeline,
so the releasing process will be blocked by just one action and not
multiple ones. This will be needed in the future, when we start
using multi-version builds.

closes: #284