Improve thread safety / parallel execution of extensions (#131 / #253)
It must be safe to use Pioneer's extensions in a test suite that is
executed in parallel. This change enables that for all extensions
except TempDirectory (see #227). It does that in various ways.

# Resource locks

For extensions touching global state (like default locales or
environment variables), we've chosen the following approach:

* the extension acquires a read/write lock to the global resource
  (this prevents extended tests from running in parallel)
* we offer a `@Writes...` annotation that does the same thing, so
  users can annotate their tests that write to the same resource and
  prevent them from running in parallel with each other and with
  extended tests
* we offer a `@Reads...` annotation that acquires read access to the
  same lock, so users can make sure such tests do not run in parallel
  with tests that write to the same resource (they can run in
  parallel with one another, though)

# Verify, then modify

Most extensions verify their configuration at some point. It helps
with writing parallel tests for them if they do not change global
state until the configuration is verified. That particularly applies
to "store in beforeEach - restore in afterEach"-extensions! If they
fail after "store", they will still "restore" and thus potentially
create a race condition with other tests.

# Testing

To have a better chance to discover threading-related problems in our
extensions, we parallelize our own tests (configured in
`junit-platform.properties`). Ideally, we'd like to run them in
parallel _across_ and _within_ top-level classes, but unfortunately,
this leads to problems[1] when some test setups change global state
(like the security manager) that other tests rely on. As we see it,
the solution would be to force such tests onto a single thread, but
Jupiter has no such feature, yet[2]. While a homegrown solution is
possible[3], we wait for the discussion to resolve. We hence do not
parallelize across top-level classes - just within.

Closes: #131
Relates to #274
PR: #253

[1]: https://github.com/junit-pioneer/junit-pioneer/pull/253#issuecomment-665235062
[2]: https://github.com/junit-team/junit5/issues/2142
[3]: https://github.com/junit-team/junit5/issues/2142#issuecomment-668409251