Temporarily disable windows builds (#324 / #342)
In #324 it is stated out that quite often
the builds on Windows time out, most probably
because of some kind of deadlock when executing
tests in parallel, not because there are errors.

When a build fails, the release workflow is
blocked, preventing us to release.

We decided to disable the Windows build temporarily
to enable releases.

related to: #324
PR: #342