Introduce `@Issue` and `IssueProcessor` extension (#135 / #361)
This PR provides the `@Issue` annotation and its extension,
which allows users to mark tests that belong to a specific
issue (e.g. requirement).

This information is not needed during test execution
(unlike, e.g., changing the time zone via `@DefaultTimeZone`
annotation), but after the test execution (e.g. in a report
to check which tests of which issues have passed/failed).
This PR also introduces the `IssueExtensionExecutionListener`.

This listener collects the name, issue ID and result of all
tests that are annotated with the introduced annotation.
After the test execution has finished, the collected
information are published to `IssueProcessor` implementations,
which are picked up via the `ServiceLoader` API.[1]

Users can implement this API for further process of the
results based on their individual needs, e.g. generating a
report, or sending mails.

With such an API, Pioneer does not makes any assumptions
about the users' needs for the information and also avoids
additional dependencies and therefore dependency conflicts.
(In fact, avoiding a dependency on JAXB for report generation
was the motivation to decouple information gathering and
processing.)

Closes: #135
PR: #361


[1] https://docs.oracle.com/javase/8/docs/api/java/util/ServiceLoader.html