Make Cartesian product tests work with range sources (#354 / #362)
A Cartesian product is the set of all ordered pairs of elements from 
at least two sets. The extension is already able to create this set
from multiple different sources, e.g. `@CartesianValueSource`.

The range source extension creates multiple parameters by taking an
interval and a step size, e.g. create every from from 1 to 10 with
step size 1. It's available for multiple variable types, including
byte, short and more.

This change combines these extensions as an easy way to provide each
combination of parameters from multiple intervals. To that end, the
range source annotations are made repeatable (only for use with
`@CartesianProductTest`) and `CartesianProductExtension` is enhanced
to use these annotations to build the product of the given intervals.

Closes: #354
PR: #362