Add `minSuccess` to RetryingTest
The `minSuccess` field specifies the minimum number of times the test must
complete without an error.

Added a `maxAttempts` field as an alias for `value` so that the intent
is clearer. Only one of `maxAttempts` or `value` can be set. This does
change the behavior slightly in that the max attempt count is no longer
required and defaults to 1. The change should be backward compatible
with existing usages though. If `maxAttempts` or `value` is not set,
the max attempts will default to `minSuccess`.