Use Gradle "versions" to find outdated dependencies (#443 / #444)
Now that dependabot is gone, we need to manually update dependencies.
The Gradle "versions" plugin[1] makes that easy. Running
`gradle dependencyUpdates` lists all dependencies for which a newer
version exists. Updates still need to be done manually (and
preferably in bulk every few weeks).

This commit extends `CONTRIBUTING.md` with a description of how to
update dependencies.

Relates to: #443
PR: #444


[1] https://github.com/ben-manes/gradle-versions-plugin