Fix annotation lookup in AbstractEntryBasedExtension (#473 / #476)
When using `findClosestEnclosingRepeatableAnnotations(...)`, outer
elements such as classes are also considered during search, if the
current element isn't annotated accordingly. This may cause an
`ExtensionConfigurationException` when an entry is both cleared (e.g.
on the class level) and set (e.g. on the method level).

This search is superfluous because Jupiter calls the extension at
this extension points already, so the annotation on outer elements
are already processed.

The solution is to just not look for annotation on enclosing elements.

Closes: #473
PR: #476