Sign artifacts only if project.version is specified
To e.g. enable JitPack usage or make `publishToMavenLocal` pass
_without_ configured signatory, signing should happen conditionally.
The given PR requires the `project.version` to be specified before
signing.

This approach is compatible with JUnit Pioneer's release build since
in this case the `project.version` is set via the environment variable
`ORG_GRADLE_PROJECT_version`, which in turn is injected as a workflow
parameter.