Delete unused exception (#500)
In #403 we moved our utility class into an internal package and
removed a reflection-based solution that copied our utility. The
reflection-based solution used `PioneerException` on failure to 
signal the user that Pioneer failed to initialise itself. Because
that class was deleted and `PioneerException` is not used elsewhere,
we can safely remove it.

PR: #500