Remove `Arguments` from `CartesianArgumentsProvider` API (#523 / #528)
Jupiter's `Arguments` interface is used for parameterized tests, where each
`Arguments` instance corresponds to one test execution and thus carries an
`Object[]` with ine argument per parameter. A `CartesianArgumentsProvider`,
on the other hand, provides a set of arguments for a single parameter and
thus there's no need to wrap them in a container. They can simply return the
arguments directly.

As a welcome side effect, this change now allows passing a arrays to a
parameter without accidentally flattening it, which is what #523 was mainly
concerned with.

Closes: #523
PR: #528