Add method-level sources to CartesianTest (#415 / #540)
Cartesian tests version 1 required users to annotate the test method
to define individual arguments. While version 2 of this extension
moves the bulk of configuration to parameter-level annotations it
should still allow method-level annotations to define the full set of
sets, for example to define them in a CSV or JSON file.

This is now possible. It's achieved by turning
`CartesianArgumentsProvider` into an internal marker interface with
two public extensions:

* `CartesianParameterArgumentsProvider`
  (for providing arguments for an individual parameter,
   i.e. takes on the former meaning of `CartesianArgumentsProvider`)
* `CartesianMethodArgumentsProvider`
  (for providing arguments for all parameters)

Related to #415
PR: #540