Allow non-static Cartesian factory methods with PER_CLASS lifecycle
JUnit allows non-static factory methods as long as the test class is
annotated with @testinstance(Lifecycle.PER_CLASS). The same is
currently not true for @CartesianTest.MethodFactory. This commit
fixes that.

The existing non-static test is unmodified, indicating that
non-static is still not allowed if the test class is not annotated
with @testinstance(Lifecycle.PER_CLASS).

PR: #628