Register Jackson modules in JacksonJsonConverter (624 / 629)
The default JacksonJsonConverter does not support classes from
java.time when using @JsonSource (and probably also @JsonFileSource
and @JsonClassPathSource) in combination with POJO method arguments.
This is the case even if JavaTimeModule is available on the class
path.

By calling findAndRegisterModules(), any module that's available on
the class path and is registered as a service, will be available for
JSON processing.

Fixes: #624
PR: #629