Improved requiredness check for static factory methods (628)
When allowing non-static factory methods in combination with
@Testinstance(Lifecycle.PER_CLASS), there would be failures if the
factory method was defined in a different class. The factory method
now must also be static if it's not defined in the test class (or a
super class of the test class).

Also added an example on how to use @CartesianTest.MethodFactory in
a nested test class.

PR: #628