Add option to suspend between retrying tests (#407 / #604)
Retrying tests may be flaky because of issues with external services.
The chance that they recover and behave correctly on the next try may
improve if there's a (short) pause between the test runs. This change
adds an option to suspend for a fixed time span (in milliseconds)
betweet tries.

The implemenation simply lets the executing thread sleep, which
isn't ideal. It would be much better if there were a way to schedule
the next execution for later and use the thread to execute other
tests in between, but Jupiter doesn't offer an API for that.

Closes: #407
PR: #604