Rename label "merge-ready" to "full-build" (#679 / #682)
Switching from the reduced, faster build to the full build of all
Java/OS/etc combinations requires adding a specific label to the PR.
It was called "merge-ready", which marks the PR as being ready to be
merged once the build is green, but that is (at least) occasionally
not the case.

To avoid confusion, we rename the label to the less leading name
"full-build".

Closes: #679
PR: #682